Add swupdate with dual-copy (and make it work)
Add dual-copy update mechanism with swupdate. Currently it works on the sd card.
It keeps two copies of the kernel, and the rootfs also. Swupdate updates the partitions
that are currently not in use, then it switches the u-boot config to boot from the newly
updated partitions. Next time it updates the other set.

Also, get rid of the partition labels in u-boot config. These are coming from the wks file,
which however isn't relevant for swupdate, which means that after update the labels are
pretty much gone.