make wwan0 work again after kernel and network stack update
wwan0 broke after walnascar bump. This commit makes it work again,
though it was an unreasonable amount of change...

Connman is out - the new network manager is systemd-networkd. Wifi is
configured with iwd, and wwan0 uses libqmi to start up.

Once the SIM is unlocked, signal-translator runs a new script:
this script pulls the IP address from the modem (though the modem mgr),
configures systemd-networkd, and starts the qmi network.

The default priority of wwan network is lower than wlan, so I would
expect that to work the way I want, out of the box.