move persistent data to /data partition, misc fixes
This commit is mainly about moving (primarily firefox) data to the
/data folder to avoid deletion, but it also comes with a number
of small changes that so far weren't committed.

1. Update libubootenv config to point to internal eMMC (fw_env.config
2. Don't rely on the fstab populated by wic, but mount /data explicitly.
The fstab created by the wks file is only present in the wic image,
but not in swu. This makes it present everywhere (base-files)
3. Move some data to the /data folder, by symlinking some $HOME/xxx
folders (set-common-files)
4. Revert swayuser's homefolder changes, back to /home. The /data
partition isn't supposed to exist when the image is built, so that
wasn't a too good idea. (sway)
5. Move the wwan-template into the pp-systemd-conf recipe, because it
makes more sense.
6. Add old, but uncommited bluealsa changes, to support systemd and ofono.