pp-common-files: remove recipe
The idea was to create symlinks in the home folder for all users
for the well known directories (especially for .mozilla), and store
the actual data in the /data folder, which is not deleted upon
update.

Now, the problem is, that when I do that, it doesn't let me log into
that user account anymore - it just refuses, without any error message.

I'm sure guess it can be configued in a way to make it accept the symlinks,
but that's for another day. For now, let's just revert this, and look
for an alternative solution.